+++ freight/tests/test_views.py
@@ -260,8 +260,8 @@ class TestSetupContractHandler(NoSocketsTestCase):
         cls.factory = RequestFactory()
 
     @patch(MODULE_PATH + ".FREIGHT_OPERATION_MODE", FREIGHT_OPERATION_MODE_MY_ALLIANCE)
-    @patch(MODULE_PATH + ".messages", autospec=True)
-    @patch(MODULE_PATH + ".tasks.run_contracts_sync", autospec=True)
+    @patch(MODULE_PATH + ".messages", spec=True)
+    @patch(MODULE_PATH + ".tasks.run_contracts_sync", spec=True)
     def test_normal(self, mock_run_contracts_sync, mock_message_plus):
         ContractHandler.objects.all().delete()
         token = Mock(spec=Token)
@@ -280,8 +280,8 @@ class TestSetupContractHandler(NoSocketsTestCase):
         self.assertEqual(response.url, reverse("freight:index"))
 
     @patch(MODULE_PATH + ".FREIGHT_OPERATION_MODE", FREIGHT_OPERATION_MODE_MY_ALLIANCE)
-    @patch(MODULE_PATH + ".messages", autospec=True)
-    @patch(MODULE_PATH + ".tasks.run_contracts_sync", autospec=True)
+    @patch(MODULE_PATH + ".messages", spec=True)
+    @patch(MODULE_PATH + ".tasks.run_contracts_sync", spec=True)
     def test_error_no_alliance_member(self, mock_run_contracts_sync, mock_message_plus):
         ContractHandler.objects.all().delete()
 
@@ -304,8 +304,8 @@ class TestSetupContractHandler(NoSocketsTestCase):
     @patch(
         MODULE_PATH + ".FREIGHT_OPERATION_MODE", FREIGHT_OPERATION_MODE_MY_CORPORATION
     )
-    @patch(MODULE_PATH + ".messages", autospec=True)
-    @patch(MODULE_PATH + ".tasks.run_contracts_sync", autospec=True)
+    @patch(MODULE_PATH + ".messages", spec=True)
+    @patch(MODULE_PATH + ".tasks.run_contracts_sync", spec=True)
     def test_error_character_not_owned(
         self, mock_run_contracts_sync, mock_message_plus
     ):
@@ -329,8 +329,8 @@ class TestSetupContractHandler(NoSocketsTestCase):
     @patch(
         MODULE_PATH + ".FREIGHT_OPERATION_MODE", FREIGHT_OPERATION_MODE_MY_CORPORATION
     )
-    @patch(MODULE_PATH + ".messages", autospec=True)
-    @patch(MODULE_PATH + ".tasks.run_contracts_sync", autospec=True)
+    @patch(MODULE_PATH + ".messages", spec=True)
+    @patch(MODULE_PATH + ".tasks.run_contracts_sync", spec=True)
     def test_error_wrong_operation_mode(
         self, mock_run_contracts_sync, mock_message_plus
     ):
@@ -482,8 +482,8 @@ class TestAddLocation(TestCase):
         cls.factory = RequestFactory()
 
     @patch(MODULE_PATH + ".FREIGHT_OPERATION_MODE", FREIGHT_OPERATION_MODE_MY_ALLIANCE)
-    @patch(MODULE_PATH + ".messages", autospec=True)
-    @patch(MODULE_PATH + ".Location.objects.update_or_create_esi", autospec=True)
+    @patch(MODULE_PATH + ".messages", spec=True)
+    @patch(MODULE_PATH + ".Location.objects.update_or_create_esi", spec=True)
     def test_normal(self, mock_update_or_create_from_esi, mock_message_plus):
         location_id = 1022167642188
         location = Location.objects.get(id=location_id)
@@ -511,8 +511,8 @@ class TestAddLocation(TestCase):
         self.assertEqual(mock_message_plus.error.call_count, 0)
 
     @patch(MODULE_PATH + ".FREIGHT_OPERATION_MODE", FREIGHT_OPERATION_MODE_MY_ALLIANCE)
-    @patch(MODULE_PATH + ".messages", autospec=True)
-    @patch(MODULE_PATH + ".Location.objects.update_or_create_esi", autospec=True)
+    @patch(MODULE_PATH + ".messages", spec=True)
+    @patch(MODULE_PATH + ".Location.objects.update_or_create_esi", spec=True)
     def test_fetching_location_fails(
         self, mock_update_or_create_from_esi, mock_message_plus
     ):